Fix bug where spawned items have a wrong position
The reason is not clear to me, but the issue popped up semi-randomly
(appeared location dependent). It seemed that the spawned item (unit)
appeared in the correct position, then 1 frame later got synced to a
nonsensical position.

It is fixed by setting the position and orientation during instantiation,
not after.

This thread may be related:
http://answers.unity3d.com/questions/934348/bug-object-instantiates-in-different-position-that.html